scripts/utils.py
import os
import psycopg2
from datetime import datetime
import logging
import time 
from dotenv import load_dotenv


dotenv_path = os.path.join(os.path.dirname(__file__), '..', '.env')
load_dotenv(dotenv_path=dotenv_path)

logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s [%(levelname)s] %(message)s',
    handlers=[
        logging.FileHandler('/opt/airflow/logs/migrasi.log'),
        logging.StreamHandler()
    ]
)
logger = logging.getLogger(__name__)

def get_db_connection():
    max_retries = 5
    retry_delay = 5
    last_exception = None
    
    for attempt in range(max_retries):
        try:
            conn = psycopg2.connect(
                dbname=os.getenv("POSTGRES_DB"),
                user=os.getenv("POSTGRES_USER"),
                password=os.getenv("POSTGRES_PASSWORD"),
                host=os.getenv("POSTGRES_HOST"),
                port=os.getenv("POSTGRES_PORT"),
                connect_timeout=10
            )
            logger.info("Database connection successful.")
            return conn
        except psycopg2.OperationalError as e:
            last_exception = e
            logger.warning(f"Connection attempt {attempt + 1} failed: {e}. Retrying in {retry_delay} seconds...")
            time.sleep(retry_delay) 
            
    raise ConnectionError("Could not connect to the database after several retries.") from last_exception

def log(message, level='info'):
    
    if level.lower() == 'error':
        logger.error(message)
    elif level.lower() == 'warning':
        logger.warning(message)
    else:
        logger.info(message)